dashes in operation IDs supported in py_client (#89)
There are specs in the wild that use dashes in the operation IDs.
While uncommon, there is no reason why code generated by py_client
should not handle these cases correctly. This patch lets the function
names be determined by the snake case of the operation ID instead of
the vanilla value.

Fixes #88.