made test discovery more obvious
The `unittest` module uses a pretty obscure logic to discover tests
(see [this StackOverflow question and answers][1]). We used
`load_tests` to specify which tests should be loaded, but it turned out
to be hard to debug and not so straightforward to follow.

A much easier way is to add `__init__.py` in all the relevant
directories so that `unittest` module discovers the files
automatically.

[1]: https://stackoverflow.com/questions/29713541/recursive-unittest-discover