Added jsonschema validation before parsing
There were a couple of issues reported since OpenAPI 3 was expected,
pushing swagger-to to fail with uninformative messages which of course
puzzled the users.

This patch introduces the validation of the spec by jsonschema so that
many issues can be timely detected and more information can be presented
to the user (*e.g.*, which JSON schema was used for validation).