Made format optional
Previously, swagger-to expected format to be properly set (a missing
value was reported as error). This is too strict as generators should
assume a sane default value and deal with integers in a sane way.

This patch makes `format` property to be optional and fixes the
generators to handle missing `format`s properly.

Fixes #104.