Removed support for Python 3.5 (#127)
Python 3.5 reached end-of-life in 2020. At least one of the dependencies
(icontract) does not support it anymore, so we had to also strip the
support for this package as well.

Note that this concerns only the environment where swagger-to is
executed. The generated code can be still assumed to run under Python
3.5.